Move code to 'features/restaurants.js'
