+++ src/components/restaurants/restaurants.js
@@ -5,11 +5,10 @@ import Restaurant from '../restaurant';
 import Tabs from '../tabs';
 import Loader from '../loader';
 import {
-  restaurantsListSelector,
-  restaurantsLoadedSelector,
-  shouldLoadRestaurantsSelector,
-} from '../../redux/selectors';
-import { loadRestaurants } from '../../redux/actions';
+  loadRestaurants, restaurantsLoadedSelector,
+  restaurantsListSelector, shouldLoadRestaurantsSelector
+}
+  from '../../redux/features/restaurants';
 
 const Restaurants = ({ restaurants, loaded, shouldLoad, loadRestaurants }) => {
   const [activeId, setActiveId] = useState(restaurants[0]?.id);
