+++ src/redux/constants.js
@@ -1,4 +1,3 @@
-export const LOAD_RESTAURANTS = 'LOAD_RESTAURANTS';
 export const LOAD_PRODUCTS = 'LOAD_PRODUCTS';
 export const LOAD_USERS = 'LOAD_USERS';
