+++ src/redux/reducer/index.js
@@ -1,5 +1,5 @@
 import order from '../features/order';
-import restaurants from './restaurants';
+import restaurants from '../features/restaurants';
 import products from './products';
 import reviews from '../features/reviews';
 import users from './users';
