+++ src/redux/selectors.js
@@ -1,21 +1,15 @@
 import { createSelector } from '@reduxjs/toolkit';
 import { orderSelector } from './features/order';
+import { restaurantSelector } from './features/restaurants';
 import { reviewSelector, reviewsSelector } from './features/reviews';
 import { isLoaded, isLoading, shouldLoad } from './utils';
 
-const restaurantsSelector = (state) => state.restaurants.entities;
 const productsSelector = (state) => state.products.entities;
 const usersSelector = (state) => state.users.entities;
 
-const restaurantsStatusSelector = (state) => state.restaurants.status;
 const productsStatusSelector = (state, props) =>
   state.products.status[props.restaurantId];
 const usersStatusSelector = (state) => state.users.status;
-
-export const restaurantsLoadedSelector = isLoaded(restaurantsStatusSelector);
-export const shouldLoadRestaurantsSelector = shouldLoad(
-  restaurantsStatusSelector
-);
 
 export const productsLoadingSelector = isLoading(productsStatusSelector);
 export const shouldLoadProductsSelector = shouldLoad(productsStatusSelector);
@@ -23,13 +17,6 @@ export const shouldLoadProductsSelector = shouldLoad(productsStatusSelector);
 export const usersLoadedSelector = isLoaded(usersStatusSelector);
 export const shouldLoadUsersSelector = shouldLoad(usersStatusSelector);
 
-export const restaurantsListSelector = createSelector(
-  restaurantsSelector,
-  Object.values
-);
-
-export const restaurantSelector = (state, { id }) =>
-  restaurantsSelector(state)[id];
 export const productSelector = (state, { id }) => productsSelector(state)[id];
 
 export const orderProductsSelector = createSelector(
